Correct split log test and use 'new' scenario
It looks like the test for a split log has gotten confused with what we expect to happen when changing the return cycle type.

So, we've corrected that element of the test, as well as using one of our recently added scenarios, which provides everything we need to confirm the split log is created correctly.
